[13] Three modules in package `src`
